requirements.txt
fastapi[all]
sqlalchemy
passlib[bcrypt]
python-jose[cryptography]